+++ source/Makefile.Objects
@@ -198,6 +198,9 @@ OBJS_DEEPKS=LCAO_deepks.o\
         orbital_precalc_k.o\
         cal_gdmx.o\
         cal_gdmx_k.o\
+        cal_gedm.o\
+        cal_gvx.o\ 
+        cal_descriptor.o\
         v_delta_precalc.o\
