+++ source/module_hamilt_lcao/module_deepks/LCAO_deepks_torch.cpp
@@ -3,13 +3,6 @@
 //as well as subroutines that prints the results for checking
 
 //The file contains 8 subroutines:
-//1. cal_descriptor : obtains descriptors which are eigenvalues of pdm
-//      by calling torch::linalg::eigh
-//2. check_descriptor : prints descriptor for checking
-//3. cal_gvx : gvx is used for training with force label, which is gradient of descriptors, 
-//      calculated by d(des)/dX = d(pdm)/dX * d(des)/d(pdm) = gdmx * gvdm
-//      using einsum
-//4. check_gvx : prints gvx into gvx.dat for checking
 //5. cal_gvepsl : gvepsl is used for training with stress label, which is derivative of
 //       descriptors wrt strain tensor, calculated by
 //       d(des)/d\epsilon_{ab} = d(pdm)/d\epsilon_{ab} * d(des)/d(pdm) = gdm_epsl * gvdm
@@ -35,353 +28,6 @@
 #include "module_hamilt_lcao/module_hcontainer/atom_pair.h"
 #include "module_parameter/parameter.h"
 
-void LCAO_Deepks::cal_descriptor_equiv(const int nat) {
-    ModuleBase::TITLE("LCAO_Deepks", "cal_descriptor_equiv");
-    ModuleBase::timer::tick("LCAO_Deepks", "cal_descriptor_equiv");
-
-    // a rather unnecessary way of writing this, but I'll do it for now
-    if (!this->d_tensor.empty()) 
-    {
-        this->d_tensor.erase(this->d_tensor.begin(), this->d_tensor.end());
-    }
-
-    for (int iat = 0; iat < nat; iat++) 
-    {
-        auto tmp = torch::zeros(des_per_atom, torch::kFloat64);
-        std::memcpy(tmp.data_ptr(), pdm[iat], sizeof(double) * tmp.numel());
-        this->d_tensor.push_back(tmp);
-    }
-
-    ModuleBase::timer::tick("LCAO_Deepks", "cal_descriptor_equiv");
-}
-
-// calculates descriptors from projected density matrices
-void LCAO_Deepks::cal_descriptor(const int nat) {
-    ModuleBase::TITLE("LCAO_Deepks", "cal_descriptor");
-    ModuleBase::timer::tick("LCAO_Deepks", "cal_descriptor");
-
-    if (GlobalV::deepks_equiv) {
-        this->cal_descriptor_equiv(nat);
-        return;
-    }
-
-    // init pdm_tensor and d_tensor
-    torch::Tensor tmp;
-
-    // if pdm_tensor and d_tensor is not empty, clear it !!
-    if (!this->d_tensor.empty()) {
-        this->d_tensor.erase(this->d_tensor.begin(), this->d_tensor.end());
-    }
-    if (!this->pdm_tensor.empty()) {
-        this->pdm_tensor.erase(this->pdm_tensor.begin(),
-                               this->pdm_tensor.end());
-    }
-
-    for (int inl = 0; inl < this->inlmax; ++inl) {
-        int nm = 2 * inl_l[inl] + 1;
-        tmp = torch::ones({nm, nm},
-                          torch::TensorOptions().dtype(torch::kFloat64));
-        for (int m1 = 0; m1 < nm; ++m1) {
-            for (int m2 = 0; m2 < nm; ++m2) {
-                tmp.index_put_({m1, m2}, this->pdm[inl][m1 * nm + m2]);
-            }
-        }
-        // torch::Tensor tmp = torch::from_blob(this->pdm[inl], { nm, nm },
-        // torch::requires_grad());
-        tmp.requires_grad_(true);
-        this->pdm_tensor.push_back(tmp);
-        this->d_tensor.push_back(torch::ones({nm}, torch::requires_grad(true)));
-    }
-
-    // cal d_tensor
-    for (int inl = 0; inl < inlmax; ++inl) {
-        torch::Tensor vd;
-        std::tuple<torch::Tensor, torch::Tensor> d_v(this->d_tensor[inl], vd);
-        // d_v = torch::symeig(pdm_tensor[inl], /*eigenvalues=*/true,
-        // /*upper=*/true);
-        d_v = torch::linalg::eigh(pdm_tensor[inl], /*uplo*/ "U");
-        d_tensor[inl] = std::get<0>(d_v);
-    }
-    ModuleBase::timer::tick("LCAO_Deepks", "cal_descriptor");
-    return;
-}
-
-void LCAO_Deepks::check_descriptor(const UnitCell& ucell) {
-    ModuleBase::TITLE("LCAO_Deepks", "check_descriptor");
-
-    if (GlobalV::MY_RANK != 0) 
-    {
-        return;
-    }
-    
-    std::ofstream ofs("descriptor.dat");
-	ofs << std::setprecision(10);
-	if (!GlobalV::deepks_equiv) 
-	{
-		for (int it = 0; it < ucell.ntype; it++) 
-		{
-			for (int ia = 0; ia < ucell.atoms[it].na; ia++) 
-			{
-				int iat = ucell.itia2iat(it, ia);
-				ofs << ucell.atoms[it].label << " atom_index " << ia + 1
-					<< " n_descriptor " << this->des_per_atom << std::endl;
-				int id = 0;
-				for (int inl = 0; inl < inlmax / ucell.nat; inl++) 
-				{
-					int nm = 2 * inl_l[inl] + 1;
-					for (int im = 0; im < nm; im++) 
-					{
-						const int ind = iat * inlmax / ucell.nat + inl;
-						ofs << d_tensor[ind].index({im}).item().toDouble()
-							<< " ";
-
-						if (id % 8 == 7) 
-						{
-							ofs << std::endl;
-						}
-						id++;
-					}
-				}
-				ofs << std::endl << std::endl;
-			}
-		}
-	} 
-	else 
-	{
-		for (int iat = 0; iat < ucell.nat; iat++) 
-		{
-			const int it = ucell.iat2it[iat];
-            ofs << ucell.atoms[it].label << " atom_index " << iat + 1
-                << " n_descriptor " << this->des_per_atom << std::endl;
-            for (int i = 0; i < this->des_per_atom; i++) 
-            {
-                ofs << this->pdm[iat][i] << " ";
-				if (i % 8 == 7) 
-				{
-					ofs << std::endl;
-				}
-			}
-            ofs << std::endl << std::endl;
-        }
-    }
-    return;
-}
-
-// calculates gradient of descriptors from gradient of projected density
-// matrices
-void LCAO_Deepks::cal_gvx(const int nat) {
-    ModuleBase::TITLE("LCAO_Deepks", "cal_gvx");
-    // preconditions
-    this->cal_gvdm(nat);
-
-    if (!gdmr_vector.empty()) 
-    {
-        gdmr_vector.erase(gdmr_vector.begin(), gdmr_vector.end());
-    }
-
-    // gdmr_vector : nat(derivative) * 3 * inl(projector) * nm * nm
-    if (GlobalV::MY_RANK == 0) 
-    {
-        // make gdmx as tensor
-        int nlmax = this->inlmax / nat;
-        for (int nl = 0; nl < nlmax; ++nl) 
-        {
-            std::vector<torch::Tensor> bmmv;
-            for (int ibt = 0; ibt < nat; ++ibt) 
-            {
-                std::vector<torch::Tensor> xmmv;
-                for (int i = 0; i < 3; ++i) 
-                {
-                    std::vector<torch::Tensor> ammv;
-                    for (int iat = 0; iat < nat; ++iat) 
-                    {
-                        int inl = iat * nlmax + nl;
-                        int nm = 2 * this->inl_l[inl] + 1;
-                        std::vector<double> mmv;
-                        for (int m1 = 0; m1 < nm; ++m1) 
-                        {
-                            for (int m2 = 0; m2 < nm; ++m2) 
-                            {
-                                if (i == 0) 
-								{
-									mmv.push_back(
-											this->gdmx[ibt][inl][m1 * nm + m2]);
-								}
-								if (i == 1) 
-								{
-									mmv.push_back(
-											this->gdmy[ibt][inl][m1 * nm + m2]);
-								}
-								if (i == 2) 
-								{
-									mmv.push_back(
-											this->gdmz[ibt][inl][m1 * nm + m2]);
-								}
-                            }
-                        } // nm^2
-                        torch::Tensor mm
-                            = torch::tensor(
-                                  mmv,
-                                  torch::TensorOptions().dtype(torch::kFloat64)).reshape({nm, nm}); // nm*nm
-                        ammv.push_back(mm);
-                    }
-                    torch::Tensor amm = torch::stack(ammv, 0); // nat*nm*nm
-                    xmmv.push_back(amm);
-                }
-                torch::Tensor bmm = torch::stack(xmmv, 0); // 3*nat*nm*nm
-                bmmv.push_back(bmm);
-            }
-            this->gdmr_vector.push_back(torch::stack(bmmv, 0)); // nbt*3*nat*nm*nm
-        }
-
-        assert(this->gdmr_vector.size() == nlmax);
-
-        // einsum for each inl:
-        // gdmr_vector : b:nat(derivative) * x:3 * a:inl(projector) * m:nm *
-        // n:nm gevdm_vector : a:inl * v:nm (descriptor) * m:nm (pdm, dim1) *
-        // n:nm (pdm, dim2) gvx_vector : b:nat(derivative) * x:3 *
-        // a:inl(projector) * m:nm(descriptor)
-        std::vector<torch::Tensor> gvx_vector;
-        for (int nl = 0; nl < nlmax; ++nl) {
-            gvx_vector.push_back(
-                at::einsum("bxamn, avmn->bxav",
-                           {this->gdmr_vector[nl], this->gevdm_vector[nl]}));
-        }
-
-        // cat nv-> \sum_nl(nv) = \sum_nl(nm_nl)=des_per_atom
-        // concatenate index a(inl) and m(nm)
-        this->gvx_tensor = torch::cat(gvx_vector, -1);
-
-        assert(this->gvx_tensor.size(0) == nat);
-        assert(this->gvx_tensor.size(1) == 3);
-        assert(this->gvx_tensor.size(2) == nat);
-        assert(this->gvx_tensor.size(3) == this->des_per_atom);
-    }
-
-    return;
-}
-
-void LCAO_Deepks::check_gvx(const int nat) {
-    std::stringstream ss;
-    std::ofstream ofs_x;
-    std::ofstream ofs_y;
-    std::ofstream ofs_z;
-
-    ofs_x << std::setprecision(12);
-    ofs_y << std::setprecision(12);
-    ofs_z << std::setprecision(12);
-
-    for (int ia = 0; ia < nat; ia++) {
-        ss.str("");
-        ss << "gvx_" << ia << ".dat";
-        ofs_x.open(ss.str().c_str());
-        ss.str("");
-        ss << "gvy_" << ia << ".dat";
-        ofs_y.open(ss.str().c_str());
-        ss.str("");
-        ss << "gvz_" << ia << ".dat";
-        ofs_z.open(ss.str().c_str());
-
-        ofs_x << std::setprecision(10);
-        ofs_y << std::setprecision(10);
-        ofs_z << std::setprecision(10);
-
-        for (int ib = 0; ib < nat; ib++) {
-            for (int inl = 0; inl < inlmax / nat; inl++) {
-                int nm = 2 * inl_l[inl] + 1;
-                {
-                    const int ind = ib * inlmax / nat + inl;
-                    ofs_x
-                        << gvx_tensor.index({ia, 0, ib, inl}).item().toDouble()
-                        << " ";
-                    ofs_y
-                        << gvx_tensor.index({ia, 1, ib, inl}).item().toDouble()
-                        << " ";
-                    ofs_z
-                        << gvx_tensor.index({ia, 2, ib, inl}).item().toDouble()
-                        << " ";
-                }
-            }
-            ofs_x << std::endl;
-            ofs_y << std::endl;
-            ofs_z << std::endl;
-        }
-        ofs_x.close();
-        ofs_y.close();
-        ofs_z.close();
-    }
-}
-
-// calculates stress of descriptors from gradient of projected density matrices
-void LCAO_Deepks::cal_gvepsl(const int nat) {
-    ModuleBase::TITLE("LCAO_Deepks", "cal_gvepsl");
-    // preconditions
-    this->cal_gvdm(nat);
-    if (!gdmepsl_vector.empty()) {
-        gdmepsl_vector.erase(gdmepsl_vector.begin(), gdmepsl_vector.end());
-    }
-
-    // gdmr_vector : nat(derivative) * 3 * inl(projector) * nm * nm
-	if (GlobalV::MY_RANK == 0) 
-	{
-		// make gdmx as tensor
-		int nlmax = this->inlmax / nat;
-		for (int nl = 0; nl < nlmax; ++nl) 
-		{
-			std::vector<torch::Tensor> bmmv;
-			for (int i = 0; i < 6; ++i) 
-			{
-				std::vector<torch::Tensor> ammv;
-				for (int iat = 0; iat < nat; ++iat) 
-				{
-                    int inl = iat * nlmax + nl;
-                    int nm = 2 * this->inl_l[inl] + 1;
-                    std::vector<double> mmv;
-                    for (int m1 = 0; m1 < nm; ++m1) {
-                        for (int m2 = 0; m2 < nm; ++m2) {
-                            mmv.push_back(this->gdm_epsl[i][inl][m1 * nm + m2]);
-                        }
-                    } // nm^2
-                    torch::Tensor mm
-                        = torch::tensor(
-                              mmv,
-                              torch::TensorOptions().dtype(torch::kFloat64))
-                              .reshape({nm, nm}); // nm*nm
-                    ammv.push_back(mm);
-                }
-                torch::Tensor bmm = torch::stack(ammv, 0); // nat*nm*nm
-                bmmv.push_back(bmm);
-            }
-            // torch::Tensor bmm = torch::stack(xmmv, 0);  //3*nat*nm*nm
-            // bmmv.push_back(bmm);
-            //}
-            this->gdmepsl_vector.push_back(
-                torch::stack(bmmv, 0)); // nbt*3*nat*nm*nm
-        }
-        assert(this->gdmepsl_vector.size() == nlmax);
-
-        // einsum for each inl:
-        // gdmepsl_vector : b:npol * a:inl(projector) * m:nm * n:nm
-        // gevdm_vector : a:inl * v:nm (descriptor) * m:nm (pdm, dim1) * n:nm
-        // (pdm, dim2) gvepsl_vector : b:npol * a:inl(projector) *
-        // m:nm(descriptor)
-        std::vector<torch::Tensor> gvepsl_vector;
-        for (int nl = 0; nl < nlmax; ++nl) {
-            gvepsl_vector.push_back(
-                at::einsum("bamn, avmn->bav",
-                           {this->gdmepsl_vector[nl], this->gevdm_vector[nl]}));
-        }
-
-        // cat nv-> \sum_nl(nv) = \sum_nl(nm_nl)=des_per_atom
-        // concatenate index a(inl) and m(nm)
-        this->gvepsl_tensor = torch::cat(gvepsl_vector, -1);
-        assert(this->gvepsl_tensor.size(0) == 6);
-        assert(this->gvepsl_tensor.size(1) == nat);
-        assert(this->gvepsl_tensor.size(2) == this->des_per_atom);
-    }
-
-    return;
-}
 
 // dDescriptor / dprojected density matrix
 void LCAO_Deepks::cal_gvdm(const int nat) {
@@ -440,153 +86,6 @@ void LCAO_Deepks::load_model(const std::string& deepks_model) {
     }
     return;
 }
-
-inline void generate_py_files(const int lmaxd, const int nmaxd) {
-
-	if (GlobalV::MY_RANK != 0) 
-	{
-		return;
-	}
-
-    std::ofstream ofs("cal_gedm.py");
-    ofs << "import torch" << std::endl;
-    ofs << "import numpy as np" << std::endl << std::endl;
-    ofs << "import sys" << std::endl;
-
-    ofs << "from deepks.scf.enn.scf import BasisInfo" << std::endl;
-    ofs << "from deepks.iterate.template_abacus import t_make_pdm" << std::endl;
-    ofs << "from deepks.utils import load_yaml" << std::endl << std::endl;
-
-    ofs << "basis = load_yaml('basis.yaml')['proj_basis']" << std::endl;
-    ofs << "model = torch.jit.load(sys.argv[1])" << std::endl;
-    ofs << "dm_eig = np.expand_dims(np.load('dm_eig.npy'),0)" << std::endl;
-    ofs << "dm_eig = torch.tensor(dm_eig, "
-           "dtype=torch.float64,requires_grad=True)"
-        << std::endl
-        << std::endl;
-
-    ofs << "dm_flat,basis_info = t_make_pdm(dm_eig,basis)" << std::endl;
-    ofs << "ec = model(dm_flat.double())" << std::endl;
-    ofs << "gedm = "
-           "torch.autograd.grad(ec,dm_eig,grad_outputs=torch.ones_like(ec))[0]"
-        << std::endl
-        << std::endl;
-
-    ofs << "np.save('ec.npy',ec.double().detach().numpy())" << std::endl;
-    ofs << "np.save('gedm.npy',gedm.double().numpy())" << std::endl;
-    ofs.close();
-
-    ofs.open("basis.yaml");
-    ofs << "proj_basis:" << std::endl;
-    for (int l = 0; l < lmaxd + 1; l++) {
-        ofs << "  - - " << l << std::endl;
-        ofs << "    - [";
-        for (int i = 0; i < nmaxd + 1; i++) {
-            ofs << "0";
-            if (i != nmaxd) {
-                ofs << ", ";
-            }
-        }
-        ofs << "]" << std::endl;
-    }
-}
-
-void LCAO_Deepks::cal_gedm_equiv(const int nat) {
-    ModuleBase::TITLE("LCAO_Deepks", "cal_gedm_equiv");
-
-	LCAO_deepks_io::save_npy_d(
-			nat, 
-			this->des_per_atom, 
-			this->inlmax, 
-			this->inl_l,
-			GlobalV::deepks_equiv, 
-			this->d_tensor, 
-			GlobalV::MY_RANK); // libnpy needed
-
-    generate_py_files(this->lmaxd, this->nmaxd);
-
-    if (GlobalV::MY_RANK == 0) {
-        std::string cmd = "python cal_gedm.py " + PARAM.inp.deepks_model;
-        int stat = std::system(cmd.c_str());
-        assert(stat == 0);
-    }
-
-    MPI_Barrier(MPI_COMM_WORLD);
-
-	LCAO_deepks_io::load_npy_gedm(
-			nat,
-			this->des_per_atom,
-			this->gedm,
-			this->E_delta,
-			GlobalV::MY_RANK);
-
-    std::string cmd = "rm -f cal_gedm.py basis.yaml";
-    std::system(cmd.c_str());
-}
-
-// obtain from the machine learning model dE_delta/dDescriptor
-void LCAO_Deepks::cal_gedm(const int nat) {
-
-    if (GlobalV::deepks_equiv) 
-    {
-        this->cal_gedm_equiv(nat);
-        return;
-    }
-
-    // using this->pdm_tensor
-    ModuleBase::TITLE("LCAO_Deepks", "cal_gedm");
-
-    // forward
-    std::vector<torch::jit::IValue> inputs;
-
-    // input_dim:(natom, des_per_atom)
-    inputs.push_back(
-        torch::cat(this->d_tensor, 0).reshape({1, nat, this->des_per_atom}));
-    std::vector<torch::Tensor> ec;
-    ec.push_back(module.forward(inputs).toTensor()); // Hartree
-    this->E_delta = ec[0].item().toDouble() * 2; // Ry; *2 is for Hartree to Ry
-
-    // cal gedm
-    std::vector<torch::Tensor> gedm_shell;
-    gedm_shell.push_back(torch::ones_like(ec[0]));
-    this->gedm_tensor = torch::autograd::grad(ec,
-                                              this->pdm_tensor,
-                                              gedm_shell,
-                                              /*retain_grad=*/true,
-                                              /*create_graph=*/false,
-                                              /*allow_unused=*/true);
-
-    // gedm_tensor(Hartree) to gedm(Ry)
-    for (int inl = 0; inl < inlmax; ++inl) {
-        int nm = 2 * inl_l[inl] + 1;
-        for (int m1 = 0; m1 < nm; ++m1) {
-            for (int m2 = 0; m2 < nm; ++m2) {
-                int index = m1 * nm + m2;
-                //*2 is for Hartree to Ry
-                this->gedm[inl][index]
-                    = this->gedm_tensor[inl].index({m1, m2}).item().toDouble()
-                      * 2;
-            }
-        }
-    }
-    return;
-}
-
-void LCAO_Deepks::check_gedm() {
-    std::ofstream ofs("gedm.dat");
-    for (int inl = 0; inl < inlmax; inl++) {
-        int nm = 2 * inl_l[inl] + 1;
-        for (int m1 = 0; m1 < nm; ++m1) {
-            for (int m2 = 0; m2 < nm; ++m2) {
-                int index = m1 * nm + m2;
-                //*2 is for Hartree to Ry
-                ofs << this->gedm[inl][index] << " ";
-            }
-        }
-        ofs << std::endl;
-    }
-}
-
 
 // prepare_psialpha and prepare_gevdm for deepks_v_delta = 2
 void LCAO_Deepks::prepare_psialpha(const int nlocal,
